requirements.txt
fastbook==0.0.29
cassandra-driver==3.28.0
python-dotenv==1.0.0
langchain==0.0.299
cassio==0.1.3
tiktoken==0.5.1
openai==0.27.7
gradio==4.3.0
